# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHIFR Messenger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  phone: string;
}

export interface AuthCodeResponse {
  message: string;
  code: string;
}

export interface VerifyInput {
  phone: string;
  code: string;
}

export interface AuthSession {
  sessionId: string;
  userId: number;
  name: string;
  phone: string;
  /** @nullable */
  avatar?: string | null;
}

export interface User {
  id: number;
  phone: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  isOnline: boolean;
  /** @nullable */
  lastSeen?: string | null;
}

export interface UpdateProfileInput {
  name?: string;
  avatar?: string;
}

export interface Message {
  id: number;
  fromUserId: number;
  toUserId: number;
  text: string;
  /** @nullable */
  encryptedText?: string | null;
  timestamp: string;
  isEncrypted: boolean;
  /** @nullable */
  readAt?: string | null;
  /** @nullable */
  fromName?: string | null;
}

export interface MessageInput {
  toUserId: number;
  text: string;
}

export interface MessageStats {
  totalSent: number;
  totalReceived: number;
  encryptedCount: number;
  contactCount: number;
}

export interface UnsubscribePushInput {
  endpoint: string;
}

export type PushSubscriptionInputKeys = {
  p256dh: string;
  auth: string;
};

export interface PushSubscriptionInput {
  endpoint: string;
  keys: PushSubscriptionInputKeys;
}

export interface AdminUser {
  id: number;
  phone: string;
  name: string;
  messageCount: number;
  createdAt: string;
}

export type GetVapidPublicKey200 = {
  publicKey: string;
};