# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SHIFR Messenger API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: messages
    description: Messaging
  - name: push
    description: Push notifications
  - name: admin
    description: Admin panel

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Send phone number to receive verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Verification code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCodeResponse"
        "400":
          description: Bad request

  /auth/verify:
    post:
      operationId: verifyCode
      tags: [auth]
      summary: Verify phone code and get session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid code

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user info
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (contacts)
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get any user's public profile
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /messages/{userId}:
    get:
      operationId: fetchMessages
      tags: [messages]
      summary: Get message history with a specific user
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized

  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to another user
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded

  /messages/stats:
    get:
      operationId: getMessageStats
      tags: [messages]
      summary: Get messaging stats for current user
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Message stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageStats"
        "401":
          description: Unauthorized

  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Register a push subscription
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionInput"
      responses:
        "200":
          description: Subscribed
        "401":
          description: Unauthorized
    delete:
      operationId: unsubscribePush
      tags: [push]
      summary: Remove a push subscription
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribePushInput"
      responses:
        "200":
          description: Unsubscribed
        "401":
          description: Unauthorized

  /push/vapid-public-key:
    get:
      operationId: getVapidPublicKey
      tags: [push]
      summary: Get VAPID public key for push subscription
      responses:
        "200":
          description: VAPID public key
          content:
            application/json:
              schema:
                type: object
                required: [publicKey]
                properties:
                  publicKey:
                    type: string

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

  /admin/messages:
    get:
      operationId: adminListMessages
      tags: [admin]
      summary: Admin - list recent messages
      responses:
        "200":
          description: Recent messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string

    AuthCodeResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string

    VerifyInput:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string

    AuthSession:
      type: object
      required: [sessionId, userId, name, phone]
      properties:
        sessionId:
          type: string
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        avatar:
          type: ["string", "null"]

    User:
      type: object
      required: [id, phone, name, isOnline]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        avatar:
          type: ["string", "null"]
        isOnline:
          type: boolean
        lastSeen:
          type: ["string", "null"]

    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string

    Message:
      type: object
      required: [id, fromUserId, toUserId, text, timestamp, isEncrypted]
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        text:
          type: string
        encryptedText:
          type: ["string", "null"]
        timestamp:
          type: string
        isEncrypted:
          type: boolean
        readAt:
          type: ["string", "null"]
        fromName:
          type: ["string", "null"]

    MessageInput:
      type: object
      required: [toUserId, text]
      properties:
        toUserId:
          type: integer
        text:
          type: string

    MessageStats:
      type: object
      required: [totalSent, totalReceived, encryptedCount, contactCount]
      properties:
        totalSent:
          type: integer
        totalReceived:
          type: integer
        encryptedCount:
          type: integer
        contactCount:
          type: integer

    UnsubscribePushInput:
      type: object
      required: [endpoint]
      properties:
        endpoint:
          type: string

    PushSubscriptionInput:
      type: object
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
            auth:
              type: string

    AdminUser:
      type: object
      required: [id, phone, name, messageCount, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string