# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHIFR Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthSession {
  sessionId: string;
  userId: number;
  name: string;
  phone: string;
  /** @nullable */
  avatar?: string | null;
}