# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHIFR Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  fromUserId: number;
  toUserId: number;
  text: string;
  /** @nullable */
  encryptedText?: string | null;
  timestamp: string;
  isEncrypted: boolean;
  /** @nullable */
  readAt?: string | null;
  /** @nullable */
  fromName?: string | null;
}