# lib/api-zod/src/generated/types/pushSubscriptionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHIFR Messenger API
 * OpenAPI spec version: 0.1.0
 */
import type { PushSubscriptionInputKeys } from "./pushSubscriptionInputKeys";

export interface PushSubscriptionInput {
  endpoint: string;
  keys: PushSubscriptionInputKeys;
}