# lib/api-zod/src/generated/types/updateProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHIFR Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileInput {
  name?: string;
  avatar?: string;
}