# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SHIFR Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  phone: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  isOnline: boolean;
  /** @nullable */
  lastSeen?: string | null;
}